Cost multipliers take ints for the correct value

+++ src/main/java/io/darkcraft/mod/common/magic/component/IDurationComponent.java
@@ -16,5 +16,5 @@ public interface IDurationComponent
 	 * Allows a spell component to alter the cost of the component based on the duration of the effect.
 	 * @return how much the component should cost with this duration
 	 */
-	public double getCostDur(double duration, double oldCost);
+	public double getCostDur(int duration, double oldCost);
 }
